move tests to another place for the injected Ray
